Changes to be committed: 	modified:   MCTDH.SRC/getparams.f90 	modified:   MCTDH.SRC/mctdhf.f90
Added capability to specify working directory on command line using
string "workdir=", in the same way that "Inp=" specifies input file.
The program will then change to workdir before reading in input files.
Default value is workdir="./" ie, the default directory to run the
program in is the launch directory.  The subroutine
getworkingdirectory() is a close copy of getinpfile().